frontend/requirements.txt
fastapi
uvicorn
langgraph
langchain-google-genai
langchain-community
tavily-python
python-dotenv
